FastAPI/blog_api/requirements.txt
alembic==1.13.3
fastapi==0.115.0
sqlalchemy-json==0.7.0
psycopg2-binary==2.9.10
uvicorn==0.31.0
email_validator